Set react-bootstrap components
